requirements.txt
discord.py
psutil
python-dotenv
aiomysql
bitcoin
bitcoinrpc
requests
python-bitcoinrpc
